Changed website for microsoft login, refresh token, added peripherals

+++ server/controllers/bookController.js
@@ -3,29 +3,52 @@ import { Book } from "../models/bookModel.js";
 import ErrorHandler from "../middlewares/errorMiddlewares.js";
 import { catchAsyncErrors } from "../middlewares/catchAsyncErrors.js";
 
-export const addBook = catchAsyncErrors(async(req, res, next)=>{
-    const {title, author, description, price, quantity} = req.body;
-    if(!title || !author || !description || !price || !quantity){
-        return next(new ErrorHandler("Please fill all fields.", 400))
-    }
-    const book = await Book.create({title, author, description, price, quantity});
-    res.status(201).json({
-        success:true,
-        message: "Book added successfully.",
-        book,
-    });
+export const addBook = catchAsyncErrors(async (req, res, next) => {
+  if (req.user.role !== "Admin") {
+    return next(new ErrorHandler("Only admins can add device models.", 403));
+  }
+
+  const {
+    name,
+    modelNumber,
+    procurementVendor,
+    manufacturer
+  } = req.body;
+
+  if (!name || !modelNumber || !procurementVendor || !manufacturer) {
+    return next(new ErrorHandler("Missing required fields.", 400));
+  }
+
+  const book = await Book.create({
+    name,
+    modelNumber,
+    procurementVendor,
+    manufacturer,
+  });
+
+  res.status(201).json({
+    success: true,
+    message: "Device model added successfully",
+    book,
+  });
 });
-export const deleteBook = catchAsyncErrors(async(req, res, next)=>{
-    const {id} = req.params;
-    const book = await Book.findById(id);
-    if(!book){
-        return next(new ErrorHandler("Book not found.", 404));
-    }
-    await book.deleteOne();
-    res.status(200).json({
-        success: true,
-        message: "Book deleted successfully.",
-    });
+
+export const deleteBook = catchAsyncErrors(async (req, res, next) => {
+  if (req.user.role !== "Admin") {
+    return next(new ErrorHandler("Only admins can delete device models.", 403));
+  }
+
+  const { id } = req.params;
+  const book = await Book.findById(id);
+  if (!book) {
+    return next(new ErrorHandler("Book not found.", 404));
+  }
+
+  await book.deleteOne();
+  res.status(200).json({
+    success: true,
+    message: "Book deleted successfully.",
+  });
 });
 export const getAllBooks = catchAsyncErrors(async(req, res, next) => {
     const books = await Book.find();
@@ -34,3 +57,4 @@ export const getAllBooks = catchAsyncErrors(async(req, res, next) => {
         books,
     });
 });
+
